Chapter Selection
Worked on the chapter selection page. You can now navigate between the `bible` page and the `chapSelect` page. Next up is doing the same thing, only for `bookSelect`. Once this is done, navigation for the Bible should be pretty much done--unless of course we want to add a search feature.